requirements.txt
fastapi
pydantic
uvicorn
jinja2
pytesseract
python-multipart
Pillow
